+++ src/Runestone.FSharp/ImportDirectory.fs
@@ -1,0 +1,10 @@
+﻿namespace Runestone.FSharp.PortableExecutable
+
+open System
+
+type ImportDirectory = {
+    ModuleName: string
+    Timestamp: DateTime
+    ImportTableAddress: uint32
+    ImportLookups: Choice<uint32 [], uint64 []>
+}