+++ src/Runestone.FSharp/OptionalHeader.fs
@@ -37,12 +37,7 @@ type DllCharacteristics =
 type DataDirectory = {
     VirtualAddress: UInt32
     Size: UInt32
-}
-    with
-        static member internal FromNative(header: IMAGE_DATA_DIRECTORY) = {
-            VirtualAddress = header.VirtualAddress
-            Size = header.Size
-        }         
+}       
 
 type OptionalHeader = {
     LinkerVersion: Version
@@ -92,93 +87,7 @@ type OptionalHeader = {
     with
         member x.Is64Bit = x.BaseOfData.IsNone
 
-        static member internal FromNative32(header: IMAGE_OPTIONAL_HEADER32) = {
-            LinkerVersion = Version(int header.MajorLinkerVersion, int header.MinorLinkerVersion)
-            OperatingSystemVersion = Version(int header.MajorOperatingSystemVersion, int header.MinorOperatingSystemVersion)
-            ImageVersion = Version(int header.MajorImageVersion, int header.MinorImageVersion)
-            SubsystemVersion = Version(int header.MajorSubsystemVersion, int header.MinorSubsystemVersion)
-            Win32VersionValue = header.Win32VersionValue
+        
 
-            SizeOfCode = header.SizeOfCode
-            SizeOfInitializedData = header.SizeOfInitializedData
-            SizeOfUnitializedData = header.SizeOfUninitializedData
-            AddressOfEntryPoint = header.AddressOfEntryPoint
-            BaseOfCode = header.BaseOfCode
-            BaseOfData = Some header.BaseOfData
-            ImageBase = uint64 header.ImageBase
-            SectionAlignment = header.SectionAlignment
-            FileAlignment = header.FileAlignment
-            SizeOfImage = header.SizeOfImage
-            SizeOfHeader = header.SizeOfHeaders
-            Checksum = header.CheckSum
-            Subsystem = enum<Subsystem> (int header.Subsystem)
-            DllCharacteristics = enum<DllCharacteristics> (int header.DllCharacteristics)
-            SizeOfStackReserve = uint64 header.SizeOfStackReserve
-            SizeOfStackCommit = uint64 header.SizeOfStackCommit
-            SizeOfHeapReserve = uint64 header.SizeOfHeapReserve
-            SizeOfheapCommit = uint64 header.SizeOfHeapReserve
-            LoaderFlags = header.LoaderFlags
-            NumberOfRvaAndSizes = header.NumberOfRvaAndSizes
 
-            ExportTable = DataDirectory.FromNative header.ExportTable
-            ImportTable = DataDirectory.FromNative header.ImportTable
-            ResourceTable = DataDirectory.FromNative header.ResourceTable
-            ExceptionTable = DataDirectory.FromNative header.ExceptionTable
-            CertificateTable = DataDirectory.FromNative header.CertificateTable
-            BaseRelocationTable = DataDirectory.FromNative header.BaseRelocationTable
-            Debug = DataDirectory.FromNative header.Debug
-            Architecture = DataDirectory.FromNative header.Architecture
-            GlobalPtr = DataDirectory.FromNative header.GlobalPtr
-            TlsTable = DataDirectory.FromNative header.TLSTable
-            LoadConfigTable = DataDirectory.FromNative header.LoadConfigTable
-            BoundImport = DataDirectory.FromNative header.BoundImport
-            ImportAddressTable = DataDirectory.FromNative header.IAT
-            DelayImportDescriptor = DataDirectory.FromNative header.DelayImportDescriptor
-            ClrRuntimeHeader = DataDirectory.FromNative header.CLRRuntimeHeader
-        }
-
-        static member internal FromNative64(header: IMAGE_OPTIONAL_HEADER64) = {
-            LinkerVersion = Version(int header.MajorLinkerVersion, int header.MinorLinkerVersion)
-            OperatingSystemVersion = Version(int header.MajorOperatingSystemVersion, int header.MinorOperatingSystemVersion)
-            ImageVersion = Version(int header.MajorImageVersion, int header.MinorImageVersion)
-            SubsystemVersion = Version(int header.MajorSubsystemVersion, int header.MinorSubsystemVersion)
-            Win32VersionValue = header.Win32VersionValue
-
-            SizeOfCode = header.SizeOfCode
-            SizeOfInitializedData = header.SizeOfInitializedData
-            SizeOfUnitializedData = header.SizeOfUninitializedData
-            AddressOfEntryPoint = header.AddressOfEntryPoint
-            BaseOfCode = header.BaseOfCode
-            BaseOfData = None
-            ImageBase = header.ImageBase
-            SectionAlignment = header.SectionAlignment
-            FileAlignment = header.FileAlignment
-            SizeOfImage = header.SizeOfImage
-            SizeOfHeader = header.SizeOfHeaders
-            Checksum = header.CheckSum
-            Subsystem = enum<Subsystem> (int header.Subsystem)
-            DllCharacteristics = enum<DllCharacteristics> (int header.DllCharacteristics)
-            SizeOfStackReserve = header.SizeOfStackReserve
-            SizeOfStackCommit = header.SizeOfStackCommit
-            SizeOfHeapReserve = header.SizeOfHeapReserve
-            SizeOfheapCommit = header.SizeOfHeapReserve
-            LoaderFlags = header.LoaderFlags
-            NumberOfRvaAndSizes = header.NumberOfRvaAndSizes
-
-            ExportTable = DataDirectory.FromNative header.ExportTable
-            ImportTable = DataDirectory.FromNative header.ImportTable
-            ResourceTable = DataDirectory.FromNative header.ResourceTable
-            ExceptionTable = DataDirectory.FromNative header.ExceptionTable
-            CertificateTable = DataDirectory.FromNative header.CertificateTable
-            BaseRelocationTable = DataDirectory.FromNative header.BaseRelocationTable
-            Debug = DataDirectory.FromNative header.Debug
-            Architecture = DataDirectory.FromNative header.Architecture
-            GlobalPtr = DataDirectory.FromNative header.GlobalPtr
-            TlsTable = DataDirectory.FromNative header.TLSTable
-            LoadConfigTable = DataDirectory.FromNative header.LoadConfigTable
-            BoundImport = DataDirectory.FromNative header.BoundImport
-            ImportAddressTable = DataDirectory.FromNative header.IAT
-            DelayImportDescriptor = DataDirectory.FromNative header.DelayImportDescriptor
-            ClrRuntimeHeader = DataDirectory.FromNative header.CLRRuntimeHeader
-        }